docs: bump the docs plugin, unify the Python docs pipeline
