+++ website/package.json
@@ -34,7 +34,7 @@
         "typescript": "5.6.3"
     },
     "dependencies": {
-        "@apify/docusaurus-plugin-typedoc-api": "^4.2.7",
+        "@apify/docusaurus-plugin-typedoc-api": "^4.3.0",
         "@apify/utilities": "^2.8.0",
         "@docusaurus/core": "^3.5.2",
         "@docusaurus/mdx-loader": "^3.5.2",
